refactor: 숫자 생성, 야구 결과 함수 파일 분리
